requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
python-socketio[asyncio_client]
fastapi-socketio
python-dotenv
mysql-connector-python
python-socketio
fastapi-socketio